Callback signature reduced and a separate method added for registering context
Signature first now defined as typedef.
The new method now registers context separately.
It is bcz the timer will be created by keypad and will also assign Callback
But context must be added by the button who starts the timer, so that callback in keypad knows which button to invoke